style: apply consistent code formatting across webapp
- Applied Prettier formatting to all webapp components
- Standardized import statements ordering
- Fixed indentation and spacing consistency
- Updated toast utilities with improved error handling
- Enhanced chain configuration formatting
- Improved Next.js config structure

+++ webapp/src/app/test-vault/page.tsx
@@ -76,9 +76,9 @@ export default function TestVaultPage() {
           abi: astaVerdeAbi.abi as any,
           functionName: "balanceOf",
           args: [address, BigInt(tokenId)],
-        });
+        }) as bigint;
         
-        if (balance > 0) {
+        if (balance > 0n) {
           addLog(`Token #${tokenId}: ${balance} owned`);
         }
       }
